requirements.txt
streamlit==1.31.0
python-dotenv==1.0.0
chromadb==0.4.22
langchain==0.1.9
langchain-community==0.0.24
langchain-openai==0.0.6
langchain-anthropic==0.1.1
langchain-google-genai==0.0.9
langchain-groq==0.0.1
openai==1.12.0
anthropic==0.18.1
google-generativeai==0.3.2
groq==0.4.1
httpx==0.27.0
onnxruntime==1.17.1
requests==2.31.0
arxiv==2.1.0
plotly==5.18.0
networkx==3.2.1
pandas==2.0.3
neo4j==5.20.0
